COMMERCE-11741 buildService

+++ modules/apps/commerce/commerce-api/src/main/java/com/liferay/commerce/model/CommerceOrderItemWrapper.java
@@ -491,7 +491,8 @@ public class CommerceOrderItemWrapper
 			setShippable(shippable);
 		}
 
-		Integer shippedQuantity = (Integer)attributes.get("shippedQuantity");
+		BigDecimal shippedQuantity = (BigDecimal)attributes.get(
+			"shippedQuantity");
 
 		if (shippedQuantity != null) {
 			setShippedQuantity(shippedQuantity);
@@ -1281,7 +1282,7 @@ public class CommerceOrderItemWrapper
 	 * @return the shipped quantity of this commerce order item
 	 */
 	@Override
-	public int getShippedQuantity() {
+	public BigDecimal getShippedQuantity() {
 		return model.getShippedQuantity();
 	}
 
@@ -2171,7 +2172,7 @@ public class CommerceOrderItemWrapper
 	 * @param shippedQuantity the shipped quantity of this commerce order item
 	 */
 	@Override
-	public void setShippedQuantity(int shippedQuantity) {
+	public void setShippedQuantity(BigDecimal shippedQuantity) {
 		model.setShippedQuantity(shippedQuantity);
 	}
